requirements.txt
blinker==1.8.2
click==8.1.7
coverage==7.6.1
exceptiongroup==1.2.2
Flask==3.0.3
gunicorn==23.0.0
importlib_metadata==8.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.1
Werkzeug==3.0.4
zipp==3.20.1
